+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Disable Start and attach gdb for wine games
 * Fix SDL_Lock/UnlockAudio that can be called multiple times (#385)
 * Remove keyboard_support setting (#386)
+* Correctly handle opening of /dev/input/event|jsdev with writing flag
 
 ## [1.4.1] - 2021-01-02
 ### Added